// src/app/Vehicle/Models/VehiclesI.ts
// Generated by https://quicktype.io

export interface VehiclesResponseI {
    Count:          number;
    Message:        string;
    SearchCriteria: string;
    Results:        Result[];
}

export interface Result {
    MakeId:          number;
    MakeName:        string;
    VehicleTypeId:   number;
    VehicleTypeName: string;
    customId: string;
}